src/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
kiwisolver==1.4.8
matplotlib==3.10.3
mysql-connector-python==9.3.0
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pycairo==1.28.0
PyGObject==3.52.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
